code/requirements.txt
cachelib==0.13.0
connexion==3.1.0
eventlet==0.36.1
Flask==3.0.3
Flask_Migrate==4.0.7
flask_openapi3==3.1.3
flask_sqlalchemy==3.1.1
psutil==6.0.0
pydantic==2.8.2
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
six==1.16.0
wmi==1.5.1
